Starting a dedicated module library for SOT packages.
There is FAR too much package name overlap in SOT packages, especially 3-pin variants, to trust any of the prepackaged modules.
The default libcms library uses something close to TO-236AB for its "SOT23", but then the SOT23GDS variant for FETs is a SOT523, which is half the width of SC59, which is slightly wider than TO-236AB.
These ones will be given specific names to differentiate them (SOT23 outline variants are needed for JEDEC TO-236AB packages and SC59, etc).

Signed-off-by: Joe Baker <jbaker@alum.wpi.edu>